--- EssayModel/temp
@@ -1,1 +1,0 @@
-temp
